refactor: implement humble pattern in EmoteAnimations plugin
